🐛 Fix issue with disabled when a filter is applied
If we search for something, say "*" and then use a filter without any hits we don't want the filters to be disabled. The way if worked, it was not possible to remove the filters that  had no hits. ThiThis whole concept of disabled might be to difficult with at lot oe edge cases. If more issues is found we it might be easier to remove it